+++ car-showroom/src/main/java/com/example/car_showroom/repository/ShowroomRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ShowroomRepository extends JpaRepository<Showroom, Integer> {
-    boolean existsByCommercialRegistrationNumber(Integer commercialRegistrationNumber);
+    boolean existsByCommercialRegistrationNumber(Long commercialRegistrationNumber);
 
 }
